Rework everything yet again; move away from header-only libraries. I'm sorry if I can't decide on how to design such a massive thing from the very beginning on my own.

+++ Math/Trigonometry.hpp
@@ -4,9 +4,9 @@
 #ifndef _DREAMYUTILITIES_INCL_MATH_TRIGONOMETRY_H
 #define _DREAMYUTILITIES_INCL_MATH_TRIGONOMETRY_H
 
-#include "../Base/Base.hpp"
+#include "../DreamyUtilitiesBase.hpp"
 
-#include "../Math/Algebra.hpp"
+#include "Algebra.hpp"
 
 namespace dreamy {
 
@@ -14,7 +14,7 @@ namespace dreamy {
 
 // Wrap angle to be between 0 and 360 degrees
 MATH_TEMP inline Type WrapAngle(Type angle) {
-  return (Type)fmod(fmod(angle, 360.0) + 360.0, 360.0);
+  return (Type)fmod(fmod((double)angle, 360.0) + 360.0, 360.0);
 };
 
 // Normalize angle to be between -180 and +180 degrees
